requirements.txt
flask==2.0.1
google-generativeai==0.1.0
python-dotenv==0.19.0